# docs/openapi.json
{
  "components" : {
    "schemas" : {
      "ApproveContentHttpRequest" : {
        "properties" : {
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "publishedAt" ],
        "type" : "object"
      },
...
      "ContentDetail" : {
        "properties" : {
          "authorDisplayName" : {
            "type" : "string"
          },
          "authorId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "categories" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "commentCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "excerpt" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "markdownSource" : {
            "type" : [ "string", "null" ]
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "renderedHtml" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "sortOrder" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "DRAFT", "APPROVED" ],
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "authorDisplayName", "categories", "commentCount", "createdAt", "id", "renderedHtml", "slug", "sortOrder", "status", "tags", "title", "updatedAt" ],
        "type" : "object"
      },
      "ContentListResponse" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "posts" : {
            "items" : {
              "$ref" : "#/components/schemas/ContentSummary"
            },
            "type" : "array"
          },
          "totalCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "page", "posts", "totalCount", "totalPages" ],
        "type" : "object"
      },
      "ContentOperationConfirmation" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "message" ],
        "type" : "object"
      },
      "ContentSummary" : {
        "properties" : {
          "authorDisplayName" : {
            "type" : "string"
          },
          "categories" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "commentCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "excerpt" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : "string"
          },
          "sortOrder" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "authorDisplayName", "categories", "commentCount", "id", "slug", "sortOrder", "tags", "title" ],
        "type" : "object"
      },
...
      "DeriveTagsHttpRequest" : {
        "properties" : {
          "existingTags" : {
            "items" : {
              "type" : "string"
            },
            "type" : [ "array", "null" ]
          },
          "markdownSource" : {
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
      "DeriveTagsResponse" : {
        "properties" : {
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "tags" ],
        "type" : "object"
      },
...
      "EditContentHttpRequest" : {
        "properties" : {
          "categoryIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : [ "array", "null" ]
          },
          "markdownSource" : {
            "type" : [ "string", "null" ]
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : [ "string", "null" ]
          },
          "sortOrder" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "summary" : {
            "type" : [ "string", "null" ]
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : [ "array", "null" ]
          },
          "title" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Bundled HTTP API for the server-streampack distribution of Streampack. Includes blog, authentication, and related site endpoints. Advisory version headers are supported: clients may send 'Accept-Version' and responses include 'Content-Version' and resolved 'Accept-Version'.",
    "title" : "server-streampack API",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/posts/pending" : {
      "get" : {
        "operationId" : "listPending",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "deleted",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentListResponse"
                }
              }
            },
            "description" : "Paginated list of draft posts"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "List pending draft posts for review",
        "tags" : [ "Admin - Post Management" ]
      }
    },
    "/admin/posts/{id}" : {
      "delete" : {
        "operationId" : "deletePost",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "hard",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentOperationConfirmation"
                }
              }
            },
            "description" : "Post deleted"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Delete a post (soft by default, hard with ?hard=true)",
        "tags" : [ "Admin - Post Management" ]
      },
      "put" : {
        "operationId" : "editPost_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditContentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post updated"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Admin edit a post",
        "tags" : [ "Admin - Post Management" ]
      }
    },
    "/admin/posts/{id}/approve" : {
      "put" : {
        "operationId" : "approvePost",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApproveContentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post approved"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Approve a draft post and set publication date",
        "tags" : [ "Admin - Post Management" ]
      }
    },
    "/admin/posts/{id}/derive-tags" : {
      "post" : {
        "operationId" : "deriveTags",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeriveTagsHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeriveTagsResponse"
                }
              }
            },
            "description" : "Derived tags"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Invalid input or AI unavailable"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Derive AI tag suggestions from unsaved editor content",
        "tags" : [ "Admin - Post Management" ]
      }
    },
...
    "/posts/{id}" : {
      "get" : {
        "description" : "Returns full post detail by its UUIDv7 identifier. Draft posts are only visible to their author and admins.",
        "operationId" : "getPostById",
        "parameters" : [ {
          "description" : "Post UUIDv7",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post detail"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "summary" : "Get a post by ID",
        "tags" : [ "Posts" ]
      },
      "put" : {
        "description" : "Updates an existing post's title, content, tags, and/or categories. Editing is admin-only. Edits re-render HTML from the updated markdown source. If a `summary` is provided, it is persisted as the post excerpt.",
        "operationId" : "editPost",
        "parameters" : [ {
          "description" : "Post UUIDv7",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditContentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post updated"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found or deleted"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Edit a post",
        "tags" : [ "Posts" ]
      }
    },
...
    "/posts/{year}/{month}/{slug}" : {
      "get" : {
        "description" : "Returns full post detail including rendered HTML, tags, categories, and comment count. Draft posts are only visible to their author and admins.",
        "operationId" : "getPostBySlug",
        "parameters" : [ {
          "description" : "Publication year",
          "example" : 2026,
          "in" : "path",
          "name" : "year",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "maximum" : 3000,
            "minimum" : 2007,
            "type" : "integer"
          }
        }, {
          "description" : "Publication month (1-12)",
          "example" : 3,
          "in" : "path",
          "name" : "month",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "maximum" : 12,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "URL-safe slug derived from the post title",
          "example" : "my-post",
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post detail"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "summary" : "Get a published post by slug",
        "tags" : [ "Posts" ]
      }
    },
...
    "/ssr/posts/{year}/{month}/{slug}" : {
      "get" : {
        "operationId" : "post",
        "parameters" : [ {
          "in" : "path",
          "name" : "year",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "month",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ssr-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local development server",
    "url" : "http://localhost:8080"
  } ]
}